requirements.txt
fastapi[all]
sqlalchemy
pydantic-settings
python-multipart